Refactor common code into FixRoomGeneratorRandomSourceBase
