+++ Zilon.Core/Zilon.Core/MapGenerators/RoomStyle/FixLargeRoomGeneratorRandomSource.cs
@@ -1,21 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-
-using JetBrains.Annotations;
 using Zilon.Core.Tactics.Spatial;
 
 namespace Zilon.Core.MapGenerators.RoomStyle
 {
-    public class FixLargeRoomGeneratorRandomSource : IRoomGeneratorRandomSource
+
+    public class FixLargeRoomGeneratorRandomSource : FixRoomGeneratorRandomSourceBase, IRoomGeneratorRandomSource
     {
-        private readonly List<Tuple<OffsetCoords, OffsetCoords>> _connections;
-
-        public FixLargeRoomGeneratorRandomSource()
+        public FixLargeRoomGeneratorRandomSource() : base()
         {
-            // 20 комнат - это 6х6 матрица
-            _connections = new List<Tuple<OffsetCoords, OffsetCoords>>(20);
-
             // Все комнаты пересекаются через всё доступное пространство.
             // Каждая комната из ряда пересекается с зеркальной комнатой.
             // Т.е. левая верхняя с правой нижней.
@@ -27,45 +21,12 @@ namespace Zilon.Core.MapGenerators.RoomStyle
                     var current = new OffsetCoords(x, y);
                     var mirror = new OffsetCoords(5 - x, 5 - y);
 
-                    _connections.Add(new Tuple<OffsetCoords, OffsetCoords>(
+                    Connections.Add(new Tuple<OffsetCoords, OffsetCoords>(
                         current,
                         mirror)
                         );
                 }
             }
-        }
-
-        /// <summary>
-        /// Выбирает комнаты, с которыми есть соединение.
-        /// </summary>
-        /// <param name="currentRoom">Текущая комната, для которой ищуются соединённые соседи.</param>
-        /// <param name="maxNeighbors">Максимальное количество соединённых соседей.</param>
-        /// <param name="availableRooms">Набор доступных для соединения соседенй.</param>
-        /// <returns>
-        /// Возвращает целевые комнаты для соединения.
-        /// </returns>
-        [NotNull, ItemNotNull]
-        public Room[] RollConnectedRooms(Room currentRoom, int maxNeighbors, IList<Room> availableRooms)
-        {
-            if (!availableRooms.Any())
-            {
-                return new Room[0];
-            }
-
-            var currentConnection = _connections.Single(x =>
-                        x.Item1.X == currentRoom.PositionX &&
-                        x.Item1.Y == currentRoom.PositionY);
-
-            var connectedRoom = availableRooms.Single(x =>
-                x.PositionX == currentConnection.Item2.X &&
-                x.PositionY == currentConnection.Item2.Y);
-
-            return new[] { connectedRoom };
-        }
-
-        public RoomInteriorObjectMeta[] RollInteriorObjects(int roomWidth, int roomHeight)
-        {
-            return new RoomInteriorObjectMeta[0];
         }
 
         /// <summary>
@@ -96,43 +57,6 @@ namespace Zilon.Core.MapGenerators.RoomStyle
         }
 
         /// <summary>
-        /// Возвращает матрицу смежности между комнатами (сеть комнат).
-        /// </summary>
-        /// <param name="rooms">Всё комнаты, которые должны быть соединены в сеть.</param>
-        /// <param name="maxNeighbors">Максимальное количество соседей у комнаты.</param>
-        /// <returns>
-        /// Возвращает словарь, представляющий собой матрицу смежности комнат.
-        /// Минимальное число соседей - 1. Максимальное - не превышает указанное в аргументе значение.
-        /// </returns>
-        public IDictionary<Room, Room[]> RollRoomNet(IEnumerable<Room> rooms, int maxNeighbors)
-        {
-            var result = new Dictionary<Room, Room[]>();
-
-            foreach (var currentRoom in rooms)
-            {
-                var currentConnection = _connections.SingleOrDefault(x =>
-                        x.Item1.X == currentRoom.PositionX &&
-                        x.Item1.Y == currentRoom.PositionY);
-
-                if (currentConnection == null)
-                {
-                    continue;
-                }
-
-                var connectedRoom = rooms.SingleOrDefault(x =>
-                    x.PositionX == currentConnection.Item2.X &&
-                    x.PositionY == currentConnection.Item2.Y);
-
-                if (connectedRoom != null)
-                {
-                    result.Add(currentRoom, new[] { connectedRoom });
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
         /// Выбрасывает случайный размер комнаты.
         /// </summary>
         /// <param name="minSize">Минимальный размер комнаты.</param>
@@ -143,30 +67,9 @@ namespace Zilon.Core.MapGenerators.RoomStyle
         /// <remarks>
         /// Источник рандома возвращает случайный размер комнаты в указанном диапазоне.
         /// </remarks>
-        private Size RollRoomSize(int minSize, int maxSize)
+        protected override Size RollRoomSize(int minSize, int maxSize)
         {
             return new Size(maxSize, maxSize);
-        }
-
-        public Size[] RollRoomSize(int minSize, int maxSize, int count)
-        {
-            var sizeList = new Size[count];
-            for (var i = 0; i < count; i++)
-            {
-                sizeList[i] = RollRoomSize(minSize, maxSize);
-            }
-
-            return sizeList;
-        }
-
-        public HexNode RollTransitionNode(IEnumerable<HexNode> openRoomNodes)
-        {
-            return openRoomNodes.First();
-        }
-
-        public IEnumerable<RoomTransition> RollTransitions(IEnumerable<RoomTransition> openTransitions)
-        {
-            return new[] { openTransitions.First() };
         }
     }
 }
