+++ Zilon.Core/Zilon.Core/Zilon.Core.csproj
@@ -124,6 +124,7 @@
     <Compile Include="MapGenerators\MapRegionHelper.cs" />
     <Compile Include="MapGenerators\MonsterGenerator.cs" />
     <Compile Include="MapGenerators\MonsterGeneratorRandomSource.cs" />
+    <Compile Include="MapGenerators\RoomStyle\FixRoomGeneratorRandomSourceBase.cs" />
     <Compile Include="MapGenerators\RoomStyle\RoomGeneratorBase.cs" />
     <Compile Include="MapGenerators\RoomStyle\RoomHelper.cs" />
     <Compile Include="MapGenerators\RoomStyle\TownSectorScheme.cs" />
